### 2020-02-04 #2067 @kthyng
Hi @will-rowe, @lpantano, and @majensen! Could one of you take over to handle this submission?

@afrubin As a non-expert in your field, I read through your paper but am a bit unsure of the package motivation. It says this is slower than other readers, but easier to understand. Do people usually use this type of package as a black box or look at the code?
