### 2020-02-04 #2067 @afrubin
> @afrubin As a non-expert in your field, I read through your paper but am a bit unsure of the package motivation. It says this is slower than other readers, but easier to understand. Do people usually use this type of package as a black box or look at the code?

@kthyng Happy to expand on this!

Bioinformatics and computational biology are expanding rapidly and ever more people from biology backgrounds (often with no programming experience) are learning to do their own analyses. I think there's value in having code that's easier to understand by newcomers and students, and I know from experience that many of them like to try to read the code as part of their learning. Also, the efficiency of reading these files is typically not a major contributor to program runtime.

The package is written in pure Python (despite compiled extensions being faster) because many of these new users aren't running a Unix-based OS and don't have a standard development environment, compilers, etc. installed, which has been a source of problems when requiring some common Python bioinformatics dependencies.
