### 2020-02-06 #2076 @whedon
[REVIEW]: fqfa: A pure Python package for genomic sequence files
**Submitting author:** @afrubin (<a href="http://orcid.org/0000-0003-1474-605X">Alan F. Rubin</a>)
**Repository:** <a href="https://github.com/CountESS-Project/fqfa/" target ="_blank">https://github.com/CountESS-Project/fqfa/</a>
**Version:** v1.0.0
