### 2020-02-07 #2076 @natir
Hello @lpantano I having some little trouble with the contents of this paper. If I clearly understand the openjournals review process I need to create issue in the paper's repository.

Sorry It's the first time I review for openjournals
