### 2020-02-10 CountESS-Project/fqfa#3 @natir
Add a benchmark to compare fqfa to other python fastx parser
Even though fqfa is written purely in python and will probably be slower than alternatives that use C modules.

It would be interesting for the reader/user to have a small benchmark, on two or three files that correspond to the fqfa usage. The reader/user can then have an idea of which impact fqfa can have on its performance compared to the alternatives.

Thanks for your work.

Just to help editor to flow this issue openjournals/joss-reviews#2076
