+++ docs/notebooks/exported/benchmarks_raw.rst
@@ -1,0 +1,249 @@
+.. code:: ipython3
+
+    import pyfastx
+    from fqfa.fastq.fastq import parse_fastq_reads
+
+Benchmark 1: list of reads
+==========================
+
+This code creates a list containing all the reads in the file. Note that
+the data structures for the reads are quite different, with two being
+package-specific objects and one being a tuple.
+
+pyfastx with index
+------------------
+
+Much of the time spent in the first example is likely spent building the
+``.fxi`` index file. This file enables direct access into the FASTQ
+file, which we are not using here. The index is quite large, much larger
+than the reads in this case:
+
+::
+
+   334M    BRCA1_input_sample.fq
+   48M     BRCA1_input_sample.fq.bz2
+   520M    BRCA1_input_sample.fq.fxi
+   68M     BRCA1_input_sample.fq.gz
+   522M    BRCA1_input_sample.fq.gz.fxi
+
+.. code:: ipython3
+
+    %time reads = [x for x in pyfastx.Fastq("BRCA1_input_sample.fq")]
+    for x in reads[:5]:
+        print(repr(x))
+    del reads
+
+
+.. parsed-literal::
+
+    CPU times: user 29.5 s, sys: 17 s, total: 46.5 s
+    Wall time: 46.5 s
+    <Read> 140313_SN743_0432_AC3TTHACXX:4:1101:5633:2224:1#0/1 with length of 16
+    <Read> 140313_SN743_0432_AC3TTHACXX:4:1101:6580:2239:1#0/1 with length of 16
+    <Read> 140313_SN743_0432_AC3TTHACXX:4:1101:6929:2242:1#0/1 with length of 16
+    <Read> 140313_SN743_0432_AC3TTHACXX:4:1101:13004:2221:1#0/1 with length of 16
+    <Read> 140313_SN743_0432_AC3TTHACXX:4:1101:14034:2219:1#0/1 with length of 16
+
+
+pyfastx without index
+---------------------
+
+This is by far the fastest for just reading data from the file, but it
+doesn’t perform any extra computation or quality value conversion.
+
+.. code:: ipython3
+
+    %time reads = [x for x in pyfastx.Fastq("BRCA1_input_sample.fq", build_index=False)]
+    for x in reads[:5]:
+        print(x)
+    del reads
+
+
+.. parsed-literal::
+
+    CPU times: user 1.68 s, sys: 384 ms, total: 2.07 s
+    Wall time: 2.07 s
+    ('140313_SN743_0432_AC3TTHACXX:4:1101:5633:2224:1#0/1', 'CCCGTGGCCTTTTCCA', 'B@CFFFFFHHHHHJJJ')
+    ('140313_SN743_0432_AC3TTHACXX:4:1101:6580:2239:1#0/1', 'TTTGGTAAAGGGTAAC', 'BBCFFDFFHHHHDHIJ')
+    ('140313_SN743_0432_AC3TTHACXX:4:1101:6929:2242:1#0/1', 'AATAATGTATGTACCT', 'BC@FFFFEFHHHHJJJ')
+    ('140313_SN743_0432_AC3TTHACXX:4:1101:13004:2221:1#0/1', 'CTATTGCGTGTGATCT', 'BCCFFFFFHHHHHJJJ')
+    ('140313_SN743_0432_AC3TTHACXX:4:1101:14034:2219:1#0/1', 'ACCCCTACCCTCTGCC', 'BBBFFFFFHHHHHJJJ')
+
+
+fqfa
+----
+
+Unlike pyfastx, fqfa takes an open file handle rather than a file name.
+In these examples, this is addressed using a context created by a with
+statement.
+
+.. code:: ipython3
+
+    with open("BRCA1_input_sample.fq") as handle:
+        %time reads = [x for x in parse_fastq_reads(handle)]
[…]
