### 2020-02-18 CountESS-Project/fqfa#3 @afrubin
Thank you for this feedback. I've added a new page to the documentation that includes some usage examples and timing information for fqfa and pyfastx.

It's available at the following branch: https://github.com/CountESS-Project/fqfa/tree/benchmark
