### 2020-02-19 fe47242
remove unused imports from bz2 example
