+++ paper.md
@@ -29,6 +29,9 @@ fqfa is a pure Python package for working with files in FASTA and FASTQ formats.
 It has no dependencies outside of the Python standard library and makes use of newer language features such as type hinting to improve readability.
 This makes it more suitable for use in notebooks or projects with simple requirements, as well as easier to understand by novice bioinformaticians and students.
 
+Although it is written in pure Python, fqfa's performance is comparable to modules using C extensions like pyfastx [@du_lmdupyfastx_2019] for tasks such as processing a FASTQ file sequentially.
+Timing results and example usage are available as part of the fqfa documentation.
+
 fqfa is released under the BSD 3-Clause License and is available from GitHub and PyPI.
 
 # Acknowledgements
