### 2020-02-19 64094a7
added examples and comparison to pyfastx
