### 2020-02-19 CountESS-Project/fqfa#3 @afrubin
Good suggestion, I've added this result to the paper and merged into master.
