### 2020-02-19 #2076 @afrubin
Thanks to @natir I've added a comparison of fqfa and pyfastx (the other currently-maintained FASTQ/FASTA Python module I could find) and included this result in the paper.
