### 2020-02-19 CountESS-Project/fqfa#7 @FlorianThibord
A couple of questions / suggestions
This is my first review for joss (https://github.com/openjournals/joss-reviews/issues/2076), and I'm not sure of where to put my comments, so I'll do it here:

The code is well written, in pure python, understandable, well documented, and maybe most importantly, this package can be easily used in a python project and does not require dependencies.

The use of modern python 3 functionalities such as f-strings, dataclass module, and type hinting sets it apart from the many implementations of fasta/fastq parsers available.

I only have a question and a suggestion:
 1- Why is there a function to write fasta records, but not a function to write fastq reads?
 2- I think pyfaidx only deals with fasta, not fastq, so it might not be relevant. Maybe mention Fastaq (https://github.com/sanger-pathogens/Fastaq) instead?

Otherwise, well done it's a great work!
