### 2020-02-19 b9c6977
added note about FASTQ output
