+++ docs/conf.py
@@ -48,7 +48,7 @@ master_doc = "index"
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
 # This pattern also affects html_static_path and html_extra_path.
-exclude_patterns = ["_build", "Thumbs.db", ".DS_Store"]
+exclude_patterns = ["_build", "Thumbs.db", ".DS_Store", "notebooks"]
 
 # Intersphinx information for documentation from other packages
 intersphinx_mapping = {"python": ("https://docs.python.org/3", None)}
