### 2020-02-19 2150fce
reviewer response (#8)
* added note about FASTQ output

* added reference for Fastaq
