+++ paper.bib
@@ -101,3 +101,17 @@
 	pmid = {3162770},
 	pages = {2444--2448}
 }
+
+@misc{hunt_sanger-pathogensfastaq_2013,
+	title = {sanger-pathogens/{Fastaq}},
+	copyright = {GPL-3.0},
+	url = {https://github.com/sanger-pathogens/Fastaq},
+	abstract = {Python3 scripts to manipulate FASTA and FASTQ files},
+	urldate = {2020-02-19},
+	publisher = {Pathogen Informatics, Wellcome Sanger Institute},
+	author = {Hunt, Martin}
+	month = sep,
+	year = {2013},
+	note = {original-date: 2013-09-06T11:54:45Z},
+	keywords = {bioinformatics, genomics, global-health, infectious-diseases, next-generation-sequencing, pathogen, research, sequencing}
+}
