### 2020-02-19 CountESS-Project/fqfa#7 @afrubin
Thank you for the thoughtful review!

> 1- Why is there a function to write fasta records, but not a function to write fastq reads?

The `FastqRead.__str__()` method writes FASTQ records, so it didn't make sense to me to add an additional function. I have added text to the documentation where the FASTQ reader functions are described to clarify this usage.

> 2- I think pyfaidx only deals with fasta, not fastq, so it might not be relevant. Maybe mention Fastaq (https://github.com/sanger-pathogens/Fastaq) instead?

Thank you for pointing out Fastaq. I've added a citation to the manuscript.
