### 2020-02-19 #2076 @afrubin
I've clarified the documentation and added a reference manuscript thanks to @FlorianThibord.
