+++ paper.bib
@@ -109,7 +109,7 @@
 	abstract = {Python3 scripts to manipulate FASTA and FASTQ files},
 	urldate = {2020-02-19},
 	publisher = {Pathogen Informatics, Wellcome Sanger Institute},
-	author = {Hunt, Martin}
+	author = {Hunt, Martin},
 	month = sep,
 	year = {2013},
 	note = {original-date: 2013-09-06T11:54:45Z},
