### 2020-02-20 5af59b2
added doctest examples to travis
