+++ .travis.yml
@@ -13,8 +13,12 @@ before_script:
   - pip3 install coverage
   - pip3 install coveralls
   - if ! $NO_MYPY; then pip3 install mypy; fi
+  - pip3 install sphinx
+  - pip3 install sphinx-rtd-theme
+  - pip3 install doctest
 script:
   - coverage run --source fqfa -m unittest
   - if ! $NO_MYPY; then mypy fqfa tests; fi
+  - cd docs && make doctest
 after_success:
   - coveralls
