### 2020-02-20 d7de3aa
removed unnecessary install doctest
