### 2020-02-20 9876a26
add documentation examples to ci (#9)
