+++ docs/conf.py
@@ -23,7 +23,7 @@ copyright = "2019-2020, Alan F Rubin"
 author = "Alan F Rubin"
 
 # The full version, including alpha/beta/rc tags
-release = "0.0.2"
+release = "1.0.0"
 
 
 # -- General configuration ---------------------------------------------------
