### 2020-02-26 #2076 @afrubin
@lpantano I have created a new release with zenodo archive that matches the paper information here: https://zenodo.org/record/3688838

I've also updated the package in pypi: https://pypi.org/project/fqfa/1.0.0/
