### 2020-02-27 #2076 @labarba
Hi folks —  I am the Associate Editor-in-Chief on rotation this week. I'm going to send this back to you for some more work, because the paper does not meet [our expectations](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain). This paper is very short: the [word counter](https://wordcounter.net) tool says 197 words. While we don't expect a full article, we ask for 250–1000 words. Our docs highlight this minimum content:

- the summary describing high-level functionality and purpose of the software for a diverse, non-specialist audience;
- a clear statement of need that illustrates the research application of the software (including references to other software meeting similar needs).

TL;DR: the article needs to be improved with more description of how the software works and how it is used in research.

Thanks!
