+++ paper.bib
@@ -115,3 +115,25 @@
 	note = {original-date: 2013-09-06T11:54:45Z},
 	keywords = {bioinformatics, genomics, global-health, infectious-diseases, next-generation-sequencing, pathogen, research, sequencing}
 }
+
+
+@misc{smith_ericvsmithdataclasses_2020,
+	title = {ericvsmith/dataclasses},
+	copyright = {Apache-2.0},
+	url = {https://github.com/ericvsmith/dataclasses},
+	abstract = {Contribute to ericvsmith/dataclasses development by creating an account on GitHub.},
+	urldate = {2020-02-27},
+	author = {Smith, Eric V.},
+	month = may,
+	year = {2017},
+	note = {original-date: 2017-05-19T20:46:27Z}
+}
+
+@conference{Kluyver:2016aa,
+	Author = {Thomas Kluyver and Benjamin Ragan-Kelley and Fernando P{\'e}rez and Brian Granger and Matthias Bussonnier and Jonathan Frederic and Kyle Kelley and Jessica Hamrick and Jason Grout and Sylvain Corlay and Paul Ivanov and Dami{\'a}n Avila and Safia Abdalla and Carol Willing},
+	Booktitle = {Positioning and Power in Academic Publishing: Players, Agents and Agendas},
+	Editor = {F. Loizides and B. Schmidt},
+	Organization = {IOS Press},
+	Pages = {87 - 90},
+	Title = {Jupyter Notebooks -- a publishing format for reproducible computational workflows},
+	Year = {2016}}
