### 2020-02-28 #2076 @afrubin
@labarba Thank you for this feedback. When I did the word count myself, I had included the acknowledgements and counted from the rendered PDF rather than the source.

I have added extra detail throughout the text and should now meet the word count requirements.

The updated zenodo record is here: https://zenodo.org/record/3690230
