### 2020-03-01 #2076 @labarba
Congratulations, @afrubin, your JOSS paper is now published! 🚀

Big thanks to our editor: @lpantano, and the reviewers: @natir, @FlorianThibord — we couldn't do this without you 🙏
